+++ IBAnimatable/SystemMoveInAnimator.swift
@@ -23,19 +23,19 @@ public class SystemMoveInAnimator: NSObject, AnimatedTransitioning {
     case .Right:
       self.transitionAnimationType = .SystemMoveIn(direction: .Right)
       self.reverseAnimationType = .SystemMoveIn(direction: .Left)
-      self.interactiveGestureType = .Pan(direction: .Left)
+      self.interactiveGestureType = .Pan(fromDirection: .Left)
     case .Top:
       self.transitionAnimationType = .SystemMoveIn(direction: .Top)
       self.reverseAnimationType = .SystemMoveIn(direction: .Bottom)
-      self.interactiveGestureType = .Pan(direction: .Bottom)
+      self.interactiveGestureType = .Pan(fromDirection: .Bottom)
     case .Bottom:
       self.transitionAnimationType = .SystemMoveIn(direction: .Bottom)
       self.reverseAnimationType = .SystemMoveIn(direction: .Top)
-      self.interactiveGestureType = .Pan(direction: .Top)
+      self.interactiveGestureType = .Pan(fromDirection: .Top)
     default:
       self.transitionAnimationType = .SystemMoveIn(direction: .Left)
       self.reverseAnimationType = .SystemMoveIn(direction: .Right)
-      self.interactiveGestureType = .Pan(direction: .Right)
+      self.interactiveGestureType = .Pan(fromDirection: .Right)
     }
     
     super.init()